Se sincroniza el municipio y se arreglan temas de variables

+++ app/Http/Controllers/InstitutionController.php
@@ -49,6 +49,7 @@ class InstitutionController extends Controller
     }
     $institution->nombre = $request->nombre;
     $institution->director = $request->director;
+    $institution->municipalitie_id = $request->municipalitie_id;
     $institution->save();
 
     return redirect('institutions');
@@ -92,17 +93,22 @@ class InstitutionController extends Controller
   public function update(Request $request, $institution)
   {
     $data = Institution::find($institution);
-    $logotipo = $request->file('logo');
+    $logotipo = $request->file('logotipo');
     $nombre_logo = $data->logotipo;
-    if (!is_null($logotipo)) {
-      unlink(public_path('img/institutions/' . $nombre_logo));
-      $ruta_destino = public_path('img/institutions/');
-      $nombre_de_archivo = time() . '.' . $logotipo->getClientOriginalExtension();
-      $logotipo->move($ruta_destino, $nombre_de_archivo);
-      $data->logotipo = $nombre_de_archivo;
+    if($nombre_logo != null){
+      if (!is_null($logotipo)) {
+        unlink(public_path('img/institutions/' . $nombre_logo));
+        $ruta_destino = public_path('img/institutions/');
+        $nombre_de_archivo = time() . '.' . $logotipo->getClientOriginalExtension();
+        $logotipo->move($ruta_destino, $nombre_de_archivo);
+        $data->logotipo = $nombre_de_archivo;
+      }
     }
     if (!is_null($request->nombre)) {
       $data->nombre = $request->nombre;
+    }
+    if (!is_null($request->director)) {
+      $data->director = $request->director;
     }
     if (!is_null($request->director)) {
       $data->director = $request->director;
